Subject: vitest requiers single threaded due to use of canvas Reason for Commit:Error: `Not implemented: HTMLCanvasElement.prototype.getContext (without installing the canvas npm package)`
Summary: When starting vitest, we receieve the above error even though the canvas package is installed. The issue below relates this to multithreading issues. The `--pool=forks` flag solves this problem as the expesnse of performance. This is a limitation of the node`canvas` package.

Notes: https://github.com/vitest-dev/vitest/issues/3025
https://github.com/vitest-dev/vitest/issues/740#issuecomment-1581232942
https://vitest.dev/guide/migration.html